requirements.txt
twitchio>=2.6.0
python-dotenv>=1.0.0
yt-dlp>=2024.0.0
pysher>=1.0.8
requests>=2.31.0
